+++ templates/dashboard.html
@@ -1,0 +1,62 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Dashboard</title>
+
+    <link
+      rel="stylesheet"
+      type="text/css"
+      href="{{ url_for('static',filename='styles/dashboard.css') }}"
+    />
+  </head>
+  <body>
+    <!-- Navbar -->
+    <nav>
+      <button>Events</button>
+      <button>Achievments</button>
+      <button>Home</button>
+      <button>Login</button>
+      <button>Timeline</button>
+      <!-- Admin Page Link-->
+      <!-- Only after login -->
+      <button>Admin</button>
+    </nav>
+    <!-- Carousel -->
+    <section>
+      <div>
+        <!-- Carousel here -->
+        <p>Placeholder for carousel</p>
+      </div>
+    </section>
+    <section>
+      <div>
+        <!-- Brief intro to our webpage/subdomain -->
+        <p>
+          Lorem ipsum dolor sit amet consectetur adipisicing elit. Eligendi
+          voluptas, incidunt accusantium neque corporis voluptatem laborum error
+          a quas rem maxime consequuntur voluptatibus sequi quisquam enim fuga
+          perspiciatis natus. Incidunt.
+        </p>
+      </div>
+    </section>
+    <!-- Grid section -->
+    <section>
+      <!-- Grid goes here -->
+      <div>
+        <h2>Clubs</h2>
+        <!-- Club boxes -->
+        <h2>Departments</h2>
+        <!-- Department boxes-->
+      </div>
+    </section>
+    <!-- Footer -->
+    <section>
+      <div>
+        <!-- RV footer here -->
+        <p>Footer</p>
+      </div>
+    </section>
+  </body>
+</html>
