+++ templates/index.html
@@ -7,6 +7,7 @@
   </head>
 
   <body>
-    <h1>DTL Project</h1>
+    <h1>DTL Project lol</h1>
+    <a href="{{ url_for('dashboard') }}"><p>dashboard</p></a>
   </body>
 </html>
